feat: widgets are now fully customizable per input
