+++ lib/flutter_media_picker.dart
@@ -10,7 +10,6 @@ export './src/service/services.dart';
 export './src/media_result.dart';
 export './src/media_picker.dart';
 export './src/enums/enums.dart';
-export './src/widgets/widgets.dart';
 
 import 'package:flutter_media_picker/flutter_media_picker.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
