+++ lib/src/inputs/input_photo.dart
@@ -8,20 +8,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter_media_picker/src/abstracts/media_picker_input.dart';
 
 import 'package:flutter_media_picker/src/media_result.dart';
-import 'package:flutter_media_picker/src/widgets/icon_button_with_text.dart';
 
 /// Input for photo used by [MediaPicker].
 class MediaPickerInputPhoto implements MediaPickerInput {
   MediaPickerInputPhoto({
     this.label = 'Photo',
-    Widget? icon,
+    this.widget,
     this.checkPageSettings,
     this.onComplete,
     this.pickFile,
-  }) : icon = icon ?? IconButtonWithText(
-          icon: Icons.camera_alt,
-          iconText: label,
-        );
+  });
 
   final Future<Uint8List?> Function()? pickFile;
 
@@ -29,7 +25,7 @@ class MediaPickerInputPhoto implements MediaPickerInput {
   String label;
 
   @override
-  Widget icon;
+  Widget? widget;
 
   @override
   Future<MediaResult> onPressed(BuildContext context) async {
