+++ lib/src/inputs/input_text.dart
@@ -6,26 +6,22 @@ import 'package:flutter/material.dart';
 import 'package:flutter_form/flutter_form.dart';
 import 'package:flutter_media_picker/src/abstracts/media_picker_input.dart';
 import 'package:flutter_media_picker/src/media_result.dart';
-import 'package:flutter_media_picker/src/widgets/icon_button_with_text.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 /// Input for text used by [MediaPicker].
 class MediaPickerInputText implements MediaPickerInput {
   MediaPickerInputText({
     this.label = 'Text',
-    Widget? icon,
+    this.widget,
     this.checkPageSettings,
     this.onComplete,
-  }) : icon = icon ?? IconButtonWithText(
-          icon: Icons.text_fields,
-          iconText: label,
-        );
+  });
 
   @override
   String label;
 
   @override
-  Widget icon;
+  Widget? widget;
 
   @override
   Future<MediaResult> onPressed(BuildContext context) async {
