+++ lib/src/media_picker.dart
@@ -104,22 +104,17 @@ import '../flutter_media_picker.dart';
 class MediaPicker extends ConsumerWidget {
   const MediaPicker({
     this.mediaPickerInputs,
-    this.header,
-    this.iconButton,
+    this.inputsDirection = Axis.horizontal,
     this.onComplete,
     this.mediaCheckPage,
-    this.buttonType = ButtonType.text,
     this.horizontalSpacing = 0,
     this.verticalSpacing = 0,
     Key? key,
   }) : super(key: key);
 
   final List<MediaPickerInput>? mediaPickerInputs;
-  final Widget Function(String label, Function onPressed)? header;
-  final Widget Function(String label, Widget icon, Function onPressed)?
-      iconButton;
   final void Function(MediaResult result)? onComplete;
-  final ButtonType buttonType;
+  final Axis inputsDirection;
   final double horizontalSpacing;
   final double verticalSpacing;
   final Widget Function(
@@ -146,61 +141,43 @@ class MediaPicker extends ConsumerWidget {
 
     return Wrap(
       alignment: WrapAlignment.center,
-      direction: Axis.horizontal,
+      direction: inputsDirection,
       spacing: horizontalSpacing,
       runSpacing: verticalSpacing,
       children: [
-        if (buttonType == ButtonType.text) ...[
-          for (final input in inputs) ...[
-            const SizedBox(height: 2.5),
-            header?.call(input.label, (BuildContext ct) async {
-                  await onPressedMediaType(ct, input);
-                }) ??
-                GestureDetector(
-                  onTap: () async {
-                    await onPressedMediaType(context, input);
-                  },
-                  child: Container(
-                    height: 55,
-                    width: MediaQuery.of(context).size.width * 0.9,
-                    decoration: const BoxDecoration(
-                      border: Border(
-                        bottom: BorderSide(
-                          color: Color(0xFF979797),
-                          width: 1,
+        for (final input in inputs) ...[
+          GestureDetector(
+            onTap: () async {
+              await onPressedMediaType(context, input);
+            },
+            child: Wrap(
+              children: [
+                input.widget ??
+                    Container(
+                      height: 55,
+                      width: MediaQuery.of(context).size.width * 0.9,
+                      decoration: const BoxDecoration(
+                        border: Border(
+                          bottom: BorderSide(
+                            color: Color(0xFF979797),
+                            width: 1,
+                          ),
+                        ),
+                      ),
+                      child: Align(
+                        alignment: Alignment.centerLeft,
+                        child: Padding(
+                          padding: const EdgeInsets.only(left: 15),
+                          child: Text(
+                            input.label,
+                            style: Theme.of(context).textTheme.headline6,
+                          ),
                         ),
                       ),
                     ),
-                    child: Align(
-                      alignment: Alignment.centerLeft,
-                      child: Padding(
-                        padding: const EdgeInsets.only(left: 15),
-                        child: Text(
-                          input.label,
-                          style: Theme.of(context).textTheme.headline6,
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-            const SizedBox(height: 2.5),
-          ],
-        ] else ...[
-          for (final input in inputs) ...[
-            iconButton?.call(input.label, input.icon, (BuildContext ct) async {
-                  await onPressedMediaType(ct, input);
-                }) ??
-                GestureDetector(
-                  onTap: () async {
-                    await onPressedMediaType(context, input);
-                  },
-                  child: Wrap(
-                    children: [
-                      input.icon,
-                    ],
-                  ),
-                ),
-          ],
+              ],
+            ),
+          ),
         ]
       ],
     );
