--- lib/src/widgets/widgets.dart
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Iconica
-//
-// SPDX-License-Identifier: BSD-3-Clause
-
-export 'icon_button_with_text.dart';
